[Allocation] adjusted nsuri of extension fixture metamodel
Note: currently the simple.fujaba file, which is located in the
algorithm.tests plugin, refers to the extension metamodel => the
algorithm.tests plugin should require the algorithm.ilp.tests plugin
=> cyclic dependency. In a perfect world we would either move the
extension metamodel to the algorithm.tests plugin (but "semantically" it
belongs to the algorithm.ilp.tests plugin) or we move the extension
specific parts of the simple.fujaba file into a new file that is located
in the algorithm.ilp.tests plugin.
For now, live with the (hidden) cyclic dep...

Note 2: the above observation is independent of this commit.